backend/requirements.txt
# JobHub Backend Requirements
# Install: pip install -r requirements.txt

Flask==3.1.2
PyJWT==2.7.0
python-dotenv==1.2.1
Werkzeug>=3.0.0
Pillow>=10.0.0
PyMySQL==1.1.0
Flask-Mail>=0.10.0
google-genai
PyPDF2>=3.0.0
python-docx>=1.0.0
